insurance-extraction-engine/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
streamlit==1.28.1
requests==2.31.0
python-multipart==0.0.6

# PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0

# OCR and NLP
paddlepaddle==2.5.2
paddleocr==2.7.0.3
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0

# Layout and document processing
layoutlm==0.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Data processing
pandas==2.1.3
pyyaml==6.0.1
jsonschema==4.20.0
python-dateutil==2.8.2

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2